Add functionality to edit and register devices
New devices can now be registered, however we still need to persist
them to the database. Currently any devices registered will be lost
when the page is refreshed.

The devices can also now be edited, however the array of devices is not
yet being updated with any edits made. I'll fix this soon. We also need
to persist any device changes made to the database.

Lastly, the device name is hard coded right now, so we need to
rescaffold the database in order to get this property, so we can remove
the hard coding