Match model validation error response when there are different errors
If the model state is invalid, an automatic HTTP 400 response is
returned with the error details. If the model state is valid, but there
are additional errors such as the given username already being taken,
then a similar response object is now returned. This will make the client
side validation easier because regardless of which object is returned,
extracting the error messages is done exactly the same.

Previous to this change, the response objects were different, which
would have made the client side validation more difficult.